requirements.txt
matplotlib
numpy
opencv-python
pillow
torch
triton
